requirements.txt
streamlit
matplotlib
py2neo
pyahocorasick
jieba
gensim
scipy
sklearn
pyltp
bokeh
pyhanlp